+++ src/decoder/encap.rs
@@ -135,8 +135,13 @@ impl<'a, 'd, U> Normal<'a, 'd, U> {
 
 impl<'a, 'd, U: unit::Unit> Normal<'a, 'd, U> {
     /// Decode the packet's ETrace payload
-    pub fn payload(self) -> Result<Payload<U::IOptions, U::DOptions>, Error> {
-        Decode::decode(self.decoder)
+    pub fn payload(self) -> Result<payload::Payload<U::IOptions, U::DOptions>, Error> {
+        let width = self.decoder.trace_type_width;
+        match self.decoder.read_bits::<u8>(width)? {
+            0 => Decode::decode(self.decoder).map(payload::Payload::InstructionTrace),
+            1 => Ok(payload::Payload::DataTrace),
+            unknown => Err(Error::UnknownTraceType(unknown)),
+        }
     }
 }
 
@@ -168,53 +173,5 @@ impl<U> PartialEq for Normal<'_, '_, U> {
 impl<'a, 'd, U> Drop for Normal<'a, 'd, U> {
     fn drop(&mut self) {
         self.decoder.reset(self.remaining);
-    }
-}
-
-/// ETrace payload of an Encapsulation [Packet]
-///
-/// This datatype represents a payload as describes in Chapter 2.1.4 and Chapter
-/// 3.3 of the Encapsulation specification.
-#[derive(Copy, Clone, Debug, PartialEq)]
-pub enum Payload<I, D> {
-    InstructionTrace(payload::Payload<I, D>),
-    DataTrace,
-}
-
-impl<I, D> Payload<I, D> {
-    /// Retrieve the encapsulated instruction trace payload
-    ///
-    /// Returns [None] if this payload is not an instruction trace payload.
-    pub fn as_instruction_trace(&self) -> Option<&payload::Payload<I, D>> {
-        match self {
-            Payload::InstructionTrace(p) => Some(p),
-            _ => None,
-        }
-    }
-}
-
-impl<I, D> From<payload::Payload<I, D>> for Payload<I, D> {
-    fn from(p: payload::Payload<I, D>) -> Self {
-        Self::InstructionTrace(p)
-    }
-}
-
-impl<I, D> TryFrom<Payload<I, D>> for payload::Payload<I, D> {
-    type Error = Payload<I, D>;
-
-    fn try_from(payload: Payload<I, D>) -> Result<Self, Self::Error> {
-        match payload {
-            Payload::InstructionTrace(p) => Ok(p),
-            p => Err(p),
-        }
-    }
-}
-
-impl<U: unit::Unit> Decode<'_, '_, U> for Payload<U::IOptions, U::DOptions> {
-    fn decode(decoder: &mut Decoder<U>) -> Result<Self, Error> {
-        match decoder.read_bits::<u8>(decoder.trace_type_width)? {
-            0 => Decode::decode(decoder).map(Self::InstructionTrace),
-            unknown => Err(Error::UnknownTraceType(unknown)),
-        }
     }
 }